Added Registrar class to manage hook registration
