+++ mobbex/upgrade/upgrade-2.7.1.php
@@ -4,5 +4,6 @@ defined('_PS_VERSION_') || exit;
 
 function upgrade_module_2_7_1($module)
 {
-    return $module->registerHooks() && $module->addExtensionHooks();
+    $registrar = new \Mobbex\Registrar();
+    return $registrar->registerHooks($module) && $registrar->addExtensionHooks();
 }
